# RFC: [Performance] C++ HeaderMap rule