--- a/src/main/java/com/google/devtools/build/lib/rules/cpp/CppConfiguration.java
+++ b/src/main/java/com/google/devtools/build/lib/rules/cpp/CppConfiguration.java
@@ -1035,6 +1035,13 @@ public final class CppConfiguration extends BuildConfiguration.Fragment {
     return cppOptions.fissionModes.contains(compilationMode);
   }
 
+  /**
+   * Returns whether we are implicitly creating header maps.
+   */
+  public boolean experimentalEnableImplicitHeaderMaps() {
+    return cppOptions.experimentalEnableImplicitHeaderMaps;
+  }
+
   /** Returns true if --build_test_dwp is set on this build. */
   public boolean buildTestDwpIsActivated() {
     return cppOptions.buildTestDwp;