--- a/src/main/java/com/google/devtools/build/lib/rules/cpp/CppOptions.java
+++ b/src/main/java/com/google/devtools/build/lib/rules/cpp/CppOptions.java
@@ -316,6 +316,18 @@ public class CppOptions extends FragmentOptions {
   public boolean processHeadersInDependencies;
 
   @Option(
+    name = "experimental_enable_implicit_headermaps",
+    defaultValue = "false",
+    category = "semantics",
+    documentationCategory = OptionDocumentationCategory.UNDOCUMENTED,
+    effectTags = {OptionEffectTag.UNKNOWN},
+    help =
+        "When building a target //a:a, generate a header map containing all of the headrs that //a:a depends "
+            + "on (if header processing is enabled for the toolchain)."
+  )
+  public boolean experimentalEnableImplicitHeaderMaps;
+
+  @Option(
     name = "copt",
     allowMultiple = true,
     defaultValue = "",