… src/main/java/com/google/devtools/build/lib/rules/cpp/HeaderMapAction.java
… src/main/java/com/google/devtools/build/lib/rules/cpp/HeaderMapInfoProvider.java
… src/main/java/com/google/devtools/build/lib/rules/objc/ObjcCommon.java
… src/main/java/com/google/devtools/build/lib/rules/objc/ObjcLibrary.java
… src/main/java/com/google/devtools/build/lib/rules/objc/ObjcRuleClasses.java
… src/test/java/com/google/devtools/build/lib/rules/cpp/ClangHeaderMapTest.java