Request for feedback on an implementation of C++ HeaderMaps in Bazel.

A HeaderMap is a data structure that allows the compiler to lookup included
headers in constant time.

Traditionally, the compiler has to parse a large string of includes, and then
search these directories for a given header. This is slow for many reasons.

The protocol of HeaderMap is implemented within compilers. Please find the
Lexer implementation in Clang.
https://clang.llvm.org/doxygen/HeaderMapTypes_8h.html
https://clang.llvm.org/doxygen/HeaderMap_8cpp_source.html

## Use case:

I'm seeing a fundamental increase in build performance by using this.

I'm using a HeaderMap for one directory with 1700 headers and a deep directory
structure. Here is the clean build time:
```
bazel-bin/Pinterest/iOS/App/PinterestDevelopment.ipa
____Elapsed time: 252.712s, Critical Path: 18.78s
```
Build time before:
```
bazel-bin/Pinterest/iOS/App/PinterestDevelopment.ipa
____Elapsed time: 373.588s, Critical Path: 18.86s
```
Additionally, this solves the problem of having namespaced headers which is used
in CocoaPods all over.

With header maps, I can alias angle bracket includes directly to the files,
which fixes: ( #306 ). This is easier to use in Bazel than `isystem` including a
directory of symlinked headers. I haven't measured the performance speedups of
that yet, but am excited to given my large number of external deps.

## Implementation:

There is a rule which allows a user to create a HeaderMap directly. This allows
a user to customize a HeaderMap which can be fed to the compiler.

The user writes a rule:
```
header_map(
   name = "MyHeaderMap"
   map = {
        # Point Header.h at it's path
        "Header.h": "Path/To/Header.h"

        # Point Header.h at it's path and add a Namespace
        "Namespace/Header.h": "Path/To/Header.h"
   }
)
```

And Bazel generates a Headermap, `MyHeaderMap.hmap`

Internally, the rule writes out a binary representation that the lexer can
interpret.

The user includes a HeaderMap in `hdrs`.

Why make it a native rule?

- It is nice to use the data structures to create the headermap directly, and
  there is no way to do this in Skylark. Passing the arguments to a helper tool
  is cumbersome.

- Eventually, I think Bazel could generate implicit headermaps without even
  having a rule ( similar to module maps ). This could be useful for some
  projects and some projects would still want to create custom ones.

TODO:
- Upstream unit tests ( The ClangHeaderMap has tests, but not yet integrated
  into Bazel )
- Add integration tests for the rule in general

This is my first PR to the Bazel repo, so any suggestions or feedback is greatly
appreciated!