@bazel-io 2017-09-09
Can one of the admins verify this patch?