@philwo 2017-09-11
@lberki @ulfjack Could you please take a look at this? Maybe Manuel would also be interested in this?