@lberki 2017-09-11
/cc @mhlopko 

First the fundamentals: how are these files passed to Clang? On the command line? What happens if two header maps conflict or if what the header maps say is inconsistent with what the outcome header search would be? How would this work with non-Clang compilers, e.g. gcc or MSVC?

Since these are important enough questions, I took the liberty of only skimming over the change, and there were two things that came to mind:
  * You seem to pass a specific string-to-string map to the new rule. It looks like it's easy to override anything the C++ compiler might want to include (e.g. `<stdio.h>`), which looks like a bit too much rope. Maybe a better approach is to generate these from the `hdrs=` declarations somehow (handwave-handwave)
  * It's true that having external utilities is a bit of an overhead, but what you get out of it is that you can change them without waiting for the next Bazel release. So all things considered, I'd prefer as much functionality as possible to be *outside* of the source of Bazel. In my experience, that makes for much less frustration.