@jerrymarino 2017-09-11
Thanks for the feedback @lberki!

> How are these files passed to Clang? On the command line?

First, a HeaderMap is included in the `hdrs` attribute, which makes it a dependency of a given rule. A user passes the file to a compiler via include options: in Bazel via the `includes` and `copts` attributes. Additionally, it needs an include opt for the current directory `-I .`

Example:

```
header_map(
    name = "SomeHeaderMap",
    map = {
        "Header.h": "Some/Header.h"
    }
)

objc_library(
    name = "Some",
    hdrs = [
        "Some/Header.h",

        # Depend on and propagate the HeaderMap
        ":SomeHeaderMap"
    ],

    # include genfiles -> SomeHeaderMap.hmap
    includes = [
        "SomeHeaderMap.hmap",

        # Needed for HeaderMaps
        "-I", "."
    ]
)

```

> What happens if two header maps conflict or if what the header maps say is inconsistent with what the outcome header search would be?

The compiler resolves conflicts internally. At a higher level, it is a similar issue if a user added conflicting include paths; Say, "Header.h" on disk in 2 different directories and both directories are `isystem` included. Similar to include order, there is a precedence order to HeaderMap includes.

> How would this work with non-Clang compilers, e.g. gcc or MSVC?

HeaderMaps are supported on [Apple's GCC Fork](https://opensource.apple.com/tarballs/gcc/).

I'm not familiar with MSVC, but it looks like on WinObjc, they have support for [HeaderMaps](https://github.com/Microsoft/WinObjC/blob/aee3a8245b48f38a9519f410cde60bfac7213895/tools/vsimporter/src/utils/hmapmaker.cpp).

> You seem to pass a specific string-to-string map to the new rule. It looks like it's easy to override anything the C++ compiler might want to include (e.g. <stdio.h>), which looks like a bit too much rope. Maybe a better approach is to generate these from the hdrs= declarations somehow (handwave-handwave)

I definitely think it should generate them based on `hdrs` ( perhaps in a similar fashion as modulemaps are ). For me, this would have netted a 2 minute decrease in clean build times, without any complexity on the my end! I can look into adding support for this, but it may be a larger change.

For some of my use cases, a custom map is required: I need to specify a custom namespace for headers. The namespace is not exposed to bazel `objc_library` rule in any way. 

For example, [in CocoaPods, headers are included from a given Pod with a namespace](https://github.com/bazelbuild/bazel/issues/306).

`<AsyncDisplayKit/Header.h>`

And on the file system:

`external/*Texture*/Arbitrary/Path/To/Header.h`

The string `AsyncDisplayKit` doesn't necessary exist at the rule or file system level, but other CocoaPods and consumers import the header with that namespace.

So for generating maps for CocoaPods or other Apple style libs, I think manually specifying namespaced headers is useful. Doing so at the `BUILD` file level seems less obtrusive and has the added benefit of being more flexible for other use cases.

> It's true that having external utilities is a bit of an overhead, but what you get out of it is that you can change them without waiting for the next Bazel release. So all things considered, I'd prefer as much functionality as possible to be outside of the source of Bazel. In my experience, that makes for much less frustration.

Waiting for official Bazel releases isn't a big issue for me. At this point, I'm using custom builds for this and other changes.