@thomasvl 2017-09-12
Drive by - Why would you want this?  Last time I looked, header maps (which were created by Xcode), they were keyed by the header name only.  So if you have `foo.h` in two directories, it couldn't model it correctly and even thought you might have the property `includes` set up to control the search order, the compiler could end up using the wrong one because of what "won" in the header map.